requirements.txt
asttokens==2.4.1
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.13.1
fsspec==2024.2.0
gymnasium==0.29.1
gymnasium[atari]
gymnasium[accept-rom-license]
ipykernel==6.29.2
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
parso==0.8.3
pettingzoo==1.24.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pygame==2.5.2
Pygments==2.17.2
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
torch==2.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
wcwidth==0.2.13
pillow~=10.2.0
langchain~=0.1.13
langchain-openai~=0.1.1
openai~=1.14.2
